registerWebViewThemeHandlers util (DH-17600)
